+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [Unreleased]
+
+## [5.2.0] - 2024-07-23
+
+### Added
+
+- added information about maximum and minimum supported Android service to `InseyeAndroidSettings`
+
+- new exception `SDKServiceToHigh` and `SDKServiceToLow` that are thrown when system service is incompatible with Unity SDK 
 
 ### Changed
 
@@ -14,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- new class [InseyeAndroidSettings](./Runtime/Android/InseyeAndroidSettings.cs) that allows changing service initialization timeout on Android device
+- new class [InseyeAndroidSettings](./Runtime/Android/InseyeAndroidSettings.cs) that allows changing service
+  initialization timeout on Android device
 
 ## [5.0.7] - 2024-04-18